% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shotdata.R
\name{download_shotdata}
\alias{download_shotdata}
\title{Function to save shot data into csv file, default in your working directory}
\usage{
download_shotdata(season, playerId, path = getwd())
}
\arguments{
\item{season}{a character string}

\item{playerId}{a character string}

\item{path}{a character string}
}
\description{
Function to save shot data into csv file, default in your working directory
}
