% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shotdata.R
\name{get_shotdata}
\alias{get_shotdata}
\title{Function to return the dataframe of a given player's shot data in a given season}
\usage{
get_shotdata(season, playerId)
}
\arguments{
\item{season}{a character string}

\item{playerId}{a character string, need to look up on nba.com}
}
\value{
dataframe
}
\description{
Function to return the dataframe of a given player's shot data in a given season
}
